server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "server" }
]

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = { version = ">=0.115.11,<0.116.0", extras = ["standard"] }
kafka-python = "^2.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
